requirements.txt
Django~=3.2.3
psycopg2~=2.8.6
djangorestframework~=3.12.4
djangorestframework-simplejwt~=4.6.0
drf-yasg~=1.20.0
uvicorn~=0.13.4
celery~=5.0.5
django-celery-results~=2.0.1
django-celery-beat~=2.2.0